LPS-65012 Move wiki-editor-configuration test classes to internal package
